+++ src/components/Common/GameRoom/index.ts
@@ -1,1 +1,2 @@
 export { default as GameRoom } from './GameRoom';
+export { default as GameRoomFetcher } from './GameRoomFetcher';
