+++ src/components/Common/index.ts
@@ -28,7 +28,7 @@ export {
 export { GameListCard } from './GameListCard';
 export { GameListCarousel } from './GameListCarousel';
 export { GamePanel } from './GamePanel';
-export { GameRoom } from './GameRoom';
+export { GameRoom, GameRoomFetcher } from './GameRoom';
 export { HomeClient } from './HomeClient';
 export { InitialNickname } from './InitialNickname';
 export { Input } from './Input';
